requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
Werkzeug==3.0.3
